backend/requirements.txt
alembic==1.15.2
click==8.1.8
colorama==0.4.6
Flask==2.2.3
flask-cors==5.0.1
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==1.4.27
typing_extensions==4.13.2
Werkzeug==2.2.3
WTForms==3.0.1
